+++ Z1R_Tracker/Z1R_WPF/WPFUI.fs
@@ -1689,7 +1689,7 @@ let makeAll(mainWindow:Window, cm:CustomComboBoxes.CanvasManager, owMapNum, hear
         data.StartingItemsAndExtras.Apply()
         // Blockers
         for i = 0 to 7 do
-            for j = 0 to 1 do
+            for j = 0 to TrackerModel.DungeonBlockersContainer.MAX_BLOCKERS_PER_DUNGEON-1 do
                 TrackerModel.DungeonBlockersContainer.SetDungeonBlocker(i,j,TrackerModel.DungeonBlocker.FromHotKeyName(data.Blockers.[i].[j]))
         if anySetProblems then
             () // TODO