+++ doc/whats-new.md
@@ -110,6 +110,7 @@ a certain item, but you forget what you set it to.
  - fixed: dungeon summary tab no longer sometime shows stale/missing info
  - fixed: fewer reminders to "consider dungeon X" when you're in dungeon X already
  - hotkeys can now be bound to any keyboard key, see [HotKeys](extras.md#hotkeys) for details
+ - you can now mark up to 3 Blockers per dungeon
  - made dungeon LEVEL/BOARD header use Zelda font
  - mouse hovering a dungeon room shows a helpful row-locator graphic in the corner, see [Row location assistance](use.md#main-dungeon-row-location) for details
  - when on the dungeon summary tab, all dungeon locations get a thick green highlight on the overworld map, to make it easy to see all dungeon locations at once